Outsource websocket functionality into useChat hook
