+++ frontend/src/page/ChatDetailsPage.tsx
@@ -1,59 +1,26 @@
 import {useParams} from "react-router-dom";
 import React, {ChangeEvent, FormEvent, useContext, useEffect, useRef, useState} from "react";
 import {AppIsLoadingContext, AppIsLoadingContextType} from "../context/AppIsLoadingContext";
-import {Chat} from "../model/Chat";
 import axios from "axios";
 import toast from "react-hot-toast";
 import FoodItemCard from "../component/FoodItemCard";
 import "./ChatDetailsPage.css";
 import ChatMessageCard from "../component/ChatMessageCard";
-import {Client} from "@stomp/stompjs";
-import {ChatMessage} from "../model/ChatMessage";
 import {UserContext, UserContextType} from "../context/UserContext";
+import useChat from "../hook/useChat";
 
 export default function ChatDetailsPage() {
     const {id} = useParams();
     const {appIsLoading, setAppIsLoading} = useContext(AppIsLoadingContext) as AppIsLoadingContextType;
-    const [chat, setChat] = useState<Chat | undefined>(undefined);
+    const {chat, setChat, sendNewMessage} = useChat(id);
     const [chatContentIsScrolled, setChatContentIsScrolled] = useState(false);
-    const [client, setClient] = useState(new Client());
     const [messageDraft, setMessageDraft] = useState("");
     const scrollRef = useRef<HTMLSpanElement>(null);
     const {redirectIfNotSignedIn} = useContext(UserContext) as UserContextType;
 
-    // TODO: In Hook oder Util auslagern!
-
     useEffect(() => {
         redirectIfNotSignedIn();
     }, [redirectIfNotSignedIn]);
-
-    useEffect(() => {
-        if (id) {
-            const chatClient = new Client();
-            chatClient.configure({
-                brokerURL: 'ws://localhost:8080/api/ws/chat',
-                onConnect: () => {
-                    chatClient.subscribe(`/topic/chat/${id}`, message => {
-                            const newMessage = JSON.parse(message.body) as ChatMessage;
-                            setChat(chat => chat && {
-                                ...chat,
-                                messages: [
-                                    ...chat.messages,
-                                    newMessage
-                                ]
-                            });
-                        }
-                    );
-                }
-            });
-            chatClient.activate();
-            setClient(chatClient);
-
-            return () => {
-                chatClient.deactivate();
-            }
-        }
-    }, [id])
 
     useEffect(() => {
         setAppIsLoading(oldValue => oldValue + 1);
@@ -66,7 +33,7 @@ export default function ChatDetailsPage() {
             .finally(() => {
                 setAppIsLoading(oldValue => Math.max(0, oldValue - 1));
             });
-    }, [id, setAppIsLoading]);
+    }, [id, setChat, setAppIsLoading]);
 
     useEffect(() => {
         scrollRef.current && scrollRef.current.scrollIntoView({behavior: 'smooth', block: 'end'});
@@ -76,9 +43,13 @@ export default function ChatDetailsPage() {
         setChatContentIsScrolled(e.currentTarget.scrollTop > 0);
     }
 
-    function sendNewMessage(e: FormEvent) {
+    function handleChangeMessageDraft(e: ChangeEvent<HTMLInputElement>) {
+        setMessageDraft(e.target.value);
+    }
+
+    function handleFormSubmit(e: FormEvent) {
         e.preventDefault();
-        client.publish({destination: `/api/ws/chat/${id}`, body: messageDraft});
+        sendNewMessage(messageDraft);
         setMessageDraft("");
     }
 
@@ -92,10 +63,6 @@ export default function ChatDetailsPage() {
         } else {
             return <></>;
         }
-    }
-
-    function handleChangeMessageDraft(e: ChangeEvent<HTMLInputElement>) {
-        setMessageDraft(e.target.value);
     }
 
     const chatCards = chat.messages.map(message => <ChatMessageCard key={message.id} message={message}/>);
@@ -112,7 +79,7 @@ export default function ChatDetailsPage() {
                 </section>
             </main>
             <footer>
-                <form className={"new-message-bar"} onSubmit={sendNewMessage}>
+                <form className={"new-message-bar"} onSubmit={handleFormSubmit}>
                     <input type={"text"} value={messageDraft} onChange={handleChangeMessageDraft}
                            placeholder={"Type your message here..."} autoFocus={true}/>
                     <button>Send</button>
